Changes from code review

+++ src/main/java/org/ohdsi/webapi/prediction/specification/LassoLogisticRegressionSettingsImpl.java
@@ -1,20 +1,29 @@
 package org.ohdsi.webapi.prediction.specification;
 
+import java.math.BigDecimal;
 import org.ohdsi.analysis.prediction.design.LassoLogisticRegressionSettings;
 
+/**
+ *
+ * @author asena5
+ */
 public class LassoLogisticRegressionSettingsImpl extends SeedSettingsImpl implements LassoLogisticRegressionSettings {
-  private Float variance = 0.01f;
+  private BigDecimal variance = BigDecimal.valueOf(0.01);
 
   /**
    * A single value used as the starting value for the automatic lambda search 
    * @return variance
    **/
   @Override
-  public Float getVariance() {
+  public BigDecimal getVariance() {
     return variance;
   }
 
-  public void setVariance(Float variance) {
+    /**
+     *
+     * @param variance
+     */
+    public void setVariance(BigDecimal variance) {
     this.variance = variance;
   }
 }
